Direct3D11开发中2，至少能跑了

+++ Engine/Engine.Graphics/FontBatch2D.cs
@@ -19,7 +19,7 @@ namespace Engine.Graphics
             TriangleVertices.AddRange(batch.TriangleVertices);
             for (int i = 0; i < batch.TriangleIndices.Count; i++)
             {
-                TriangleIndices.Add((ushort)(batch.TriangleIndices[i] + count));
+                TriangleIndices.Add(batch.TriangleIndices[i] + count);
             }
             if (matrix.HasValue && matrix != Matrix.Identity)
             {
@@ -118,12 +118,12 @@ namespace Engine.Graphics
 					TriangleVertices.Array[count + 3] = new VertexPositionColorTexture(new Vector3(vector7.X, vector7.Y, depth), color, new Vector2(glyph.TexCoord1.X, glyph.TexCoord2.Y));
 					int count2 = TriangleIndices.Count;
 					TriangleIndices.Count += 6;
-					TriangleIndices.Array[count2] = (ushort)count;
-					TriangleIndices.Array[count2 + 1] = (ushort)(count + 1);
-					TriangleIndices.Array[count2 + 2] = (ushort)(count + 2);
-					TriangleIndices.Array[count2 + 3] = (ushort)(count + 2);
-					TriangleIndices.Array[count2 + 4] = (ushort)(count + 3);
-					TriangleIndices.Array[count2 + 5] = (ushort)count;
+					TriangleIndices.Array[count2] = count;
+					TriangleIndices.Array[count2 + 1] = count + 1;
+					TriangleIndices.Array[count2 + 2] = count + 2;
+					TriangleIndices.Array[count2 + 3] = count + 2;
+					TriangleIndices.Array[count2 + 4] = count + 3;
+					TriangleIndices.Array[count2 + 5] =count;
 				}
 				float num = ((i < text.Length - 1) ? base.Font.GetKerning(c, text[i + 1]) : 0f);
 				vector += right * (glyph.Width - num + fullSpacing);
